+++ src/X/Util/Cipher.php
@@ -8,6 +8,7 @@
  */
 namespace X\Util;
 use \X\Util\Loader;
+use \X\Util\Logger;
 
 final class Cipher {
 
@@ -20,8 +21,10 @@ final class Cipher {
    * @return string
    */
   public static function encode_sha256(string $plaintext, string $key = null): string {
-    if (empty($key)) $key = Loader::config('config', 'encryption_key');
-    if (empty($key)) throw new \RuntimeException('Cant find encryption_key in application/config/config.php file');
+    if (empty($key))
+      $key = Loader::config('config', 'encryption_key');
+    if (empty($key))
+      throw new \RuntimeException('Cant find encryption_key in application/config/config.php file');
     return hash('sha256', $plaintext . $key);
   }
 
@@ -114,16 +117,94 @@ final class Cipher {
   }
 
   /**
-   * @deprecated deprecated since version 3.1.0
+   * Generate key pair.
+   *
+   * ```php
+   * use \X\Util\Cipher;
+   * 
+   * // Generate 4096bit long RSA key pair.
+   * Cipher::generateKeyPair($privKey, $pubKey, [
+   *   'digest_alg' => 'sha512',
+   *   'private_key_bits' => 4096,
+   *   'private_key_type' => OPENSSL_KEYTYPE_RSA
+   * ]);
+   *
+   * // Debug private key.
+   * echo $privKey;
+   *
+   * // Debug public key.
+   * echo $pubKey;
+   *
+   * // OpenSSH encode the public key.
+   * $pubKey = Cipher::encodeOpenSshPublicKey($privKey);
+   * 
+   * // Debug OpenSSH-encoded public key.
+   * echo $pubKey;
+   * ```
+   * 
+   * @param  string &$privKey                    The generated private key is set.
+   * @param  string &$pubKey                     The generated public key is set.
+   * @param  string $options[digest_alg]         Digest method or signature hash, usually one of openssl_get_md_methods().
+   *                                             The default value is "sha512".
+   * @param  string $options[x509_extensions]    Selects which extensions should be used when creating an x509 certificate.
+   *                                             The default value is none.
+   * @param  string $options[req_extensions]     Selects which extensions should be used when creating a CSR.
+   *                                             The default value is none.
+   * @param  int    $options[private_key_bits]   Specifies how many bits should be used to generate a private key.
+   *                                             The default value is 4096.
+   * @param  int    $options[private_key_type]   Specifies the type of private key to create. This can be one of OPENSSL_KEYTYPE_DSA, OPENSSL_KEYTYPE_DH, OPENSSL_KEYTYPE_RSA or OPENSSL_KEYTYPE_EC.
+   *                                             The default value is OPENSSL_KEYTYPE_RSA.
+   * @param  bool   $options[encrypt_key]        Should an exported key (with passphrase) be encrypted?
+   * @param  int    $options[encrypt_key_cipher] One of cipher constants.
+   *                                             The default value is none.
+   * @param  string $options[curve_name]         One of openssl_get_curve_names().
+   *                                             The default value is none.
+   * @param  string $options[config]             Path to your own alternative openssl.conf file.
+   *                                             The default value is none.
    */
-  public static function encode(string $plaintext): string {
-    return self::encrypt($plaintext);
+  public static function generateKeyPair(&$privKey, &$pubKey, array $options = []) {
+    // Init options.
+    $options = array_merge([
+      'digest_alg' => 'sha512',
+      'private_key_bits' => 4096,
+      'private_key_type' => OPENSSL_KEYTYPE_RSA
+    ], $options);
+
+    // Create the private and public key
+    $privKeyResource = openssl_pkey_new($options);
+
+    // Extract the private key from $res to $privKey
+    openssl_pkey_export($privKeyResource, $privKey);
+
+    // Extract the public key from $res to $pubKey
+    $pubKey = openssl_pkey_get_details($privKeyResource)['key'];
   }
 
   /**
-   * @deprecated deprecated since version 3.1.0
+   * OpenSSH encoding the public meeting.
+   * 
+   * @param  string $privKey Public key content
+   * @return string          SSH-encoded public key
    */
-  public static function decode(string $encrypted): string {
-    return self::decrypt($encrypted);
+  public static function encodeOpenSshPublicKey(string $privKey): string {
+    $privKeyResource = openssl_pkey_get_private($privKey);
+    $keyInfo = openssl_pkey_get_details($privKeyResource);
+    $buffer  = pack('N', 7) . 'ssh-rsa' . self::encodeOpenSshBuffer($keyInfo['rsa']['e']) . self::encodeOpenSshBuffer($keyInfo['rsa']['n']);
+    return 'ssh-rsa ' . base64_encode($buffer);
+  }
+
+  /**
+   * OpenSSH encode the buffer.
+   * 
+   * @param  string $buffer buffer
+   * @return string         SSH encoded buffer
+   */
+  private static function encodeOpenSshBuffer(string $buffer): string {
+    $len = strlen($buffer);
+    if (ord($buffer[0]) & 0x80) {
+      $len++;
+      $buffer = "\x00" . $buffer;
+    }
+    return pack('Na*', $len, $buffer);
   }
 }